Create explicit HealthCheck request class (4 optional value type), remove un-necessary JsonPropertyAttributes from response class. Rename type->Type
