+++ DigitalOcean.API/Models/Requests/LoadBalancer.cs
@@ -1,8 +1,5 @@
-﻿using DigitalOcean.API.Models.Responses;
+﻿using System.Collections.Generic;
 using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace DigitalOcean.API.Models.Requests {
 	public class LoadBalancer {
@@ -43,7 +40,7 @@ namespace DigitalOcean.API.Models.Requests {
 		/// An object specifying sticky sessions settings for the Load Balancer.
 		/// </summary>
 		[JsonProperty("sticky_sessions")]
-		public StickySessions StickySessions { get; set; }
+		public Models.Responses.StickySessions StickySessions { get; set; }
 
 		/// <summary>
 		/// A boolean value indicating whether HTTP requests to the Load Balancer on port 80
